# subscriptions_backend/swagger.yaml
openapi: 3.0.0
info:
  title: Subscription Tracker API
  description: API for tracking and validating subscription usages
  version: 1.0.0
servers:
  - url: http://localhost:5001
    description: Local Development Server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
    
    Subscription:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        cost:
          type: number
        category:
          type: string
        billing_cycle:
          type: string
          enum: [monthly, yearly]
        status:
          type: string
          enum: [active, cancelled, paused]

paths:
  /api/auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
      responses:
        '201':
          description: User created
  
  /api/auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string

  /api/subscriptions:
    get:
      summary: List all subscriptions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subscription'
    post:
      summary: Create a subscription
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                cost:
                  type: number
                category:
                  type: string
      responses:
        '201':
          description: Subscription created

  /api/usage:
    get:
      summary: Get usage logs
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of usage logs

  /api/analytics/dashboard:
    get:
      summary: Get analytics dashboard
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dashboard data